business/business-antenna-connector/src/main/java/eu/estcube/gsconnector/commands/radio/GetFunc.java
package eu.estcube.gsconnector.commands.radio;

import eu.estcube.domain.TelemetryCommand;
import eu.estcube.gsconnector.commands.HamlibCommandBuilder;

public class GetFunc implements HamlibCommandBuilder {

    public StringBuilder createMessageString(TelemetryCommand command) {
        StringBuilder messageString = new StringBuilder();
        messageString.append("+u ");
        messageString.append(command.getParameter("Func"));
        messageString.append("\n");
        return messageString;
    }
}
